(minor) replace img-pan-zoom with haunted-pan-zoom
