+++ demo/index.html
@@ -7,6 +7,7 @@
 		<title>cosmoz-image-viewer demo</title>
 
 		<script src="/components/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script src="/components/resize-observer-polyfill/dist/ResizeObserver.js"></script>
 
 		<script type="module" src="/components/@polymer/iron-demo-helpers/demo-pages-shared-styles.js"></script>
 		<script type="module" src="/components/@polymer/iron-demo-helpers/demo-snippet.js"></script>
