+++ lib/directives/ref.js
@@ -1,0 +1,5 @@
+import { directive } from 'lit-html';
+
+export const ref = directive(ref => part => {
+	ref.current = part.startNode.parentNode;
+});
