+++ lib/haunted-pan-zoom.js
@@ -1,0 +1,69 @@
+/* eslint-disable multiline-ternary */
+import { component, useEffect, useCallback } from 'haunted';
+import { html, nothing } from 'lit-html';
+import { styleMap } from 'lit-html/directives/style-map';
+import { usePublicAPI } from './hooks/use-public-api';
+import { usePanZoom, ERROR } from './hooks/use-pan-zoom';
+import { ref } from './directives/ref';
+import { useDispatchEvent } from './hooks/use-dispatch-event';
+
+const HauntedPanZoom = ({ src, disabled }) => {
+	const {
+		status, zoom,
+		load, error, zoomTo, resize,
+		startMousePan, startTouchPan, startPinch,
+		onImageLoad, onWheel,
+		style,
+		containerRef
+	} = usePanZoom();
+
+	useEffect(load, [src]);
+	usePublicAPI(() => ({ zoomTo, resize }), [zoomTo, resize]);
+	usePublicAPI(() => ({ zoom }), [zoom]);
+
+	if (!src) {
+		return nothing;
+	}
+
+	const dispatchEvent = useDispatchEvent();
+	useEffect(() => dispatchEvent(new CustomEvent('zoom-changed', {detail: {value: zoom}})), [dispatchEvent, zoom]);
+	useEffect(() => dispatchEvent(new CustomEvent('status-changed', {detail: {value: status}})), [dispatchEvent, status]);
+
+	const onTouchStart = useCallback(ev => {
+		startTouchPan(ev);
+		startPinch(ev);
+	}, [startTouchPan, startPinch]);
+
+	return html`
+		<style>
+			:host {
+				display: block;
+			}
+
+			.container {
+				position: relative;
+				width: 100%;
+				height: 100%;
+				overflow: hidden;
+				touch-action: none;
+				user-select: none;
+			}
+		</style>
+		<div class="container ${status}"
+			style="pointer-events: ${disabled ? 'none' : 'all'}"
+			@wheel=${onWheel}>
+			${ref(containerRef)}
+			${!src || status === ERROR ? nothing : html`
+				<img src="${src}" draggable="false"
+					@mousedown=${startMousePan}
+					@touchstart=${onTouchStart}
+					@load=${onImageLoad} @error=${error}
+					style=${styleMap(style)}/>
+			`}
+		</div>
+	`;
+};
+
+HauntedPanZoom.observedAttributes = ['src', 'disabled'];
+
+customElements.define('haunted-pan-zoom', component(HauntedPanZoom));
