+++ package.json
@@ -32,7 +32,6 @@
 		"@fabricelements/skeleton-carousel": "^3.0.2",
 		"@neovici/cosmoz-i18next": "github:neovici/cosmoz-i18next#semver:^3.0.1",
 		"@neovici/cosmoz-utils": "github:neovici/cosmoz-utils#semver:^3.0.4",
-		"@neovici/img-pan-zoom": "github:neovici/img-pan-zoom#semver:^3.0.1",
 		"@neovici/nullxlsx": "github:neovici/nullxlsx#semver:^3.0.1",
 		"@polymer/iron-flex-layout": "^3.0.0",
 		"@polymer/iron-icons": "^3.0.0",
@@ -41,7 +40,9 @@
 		"@polymer/iron-resizable-behavior": "^3.0.0",
 		"@polymer/paper-icon-button": "^3.0.0",
 		"@polymer/polymer": "^3.0.0",
-		"@webcomponents/shadycss": "^1.9.1"
+		"@webcomponents/shadycss": "^1.9.1",
+		"haunted": "^4.5.1",
+		"lit-html": "^1.1.2"
 	},
 	"devDependencies": {
 		"@neovici/eslint-config": "github:neovici/eslint-config#semver:^1.0.0",
@@ -58,6 +59,7 @@
 		"eslint-plugin-mocha": "^5.0.0",
 		"mocha": "^5.0.0",
 		"polymer-cli": "^1.9.4",
+		"resize-observer-polyfill": "^1.5.1",
 		"sinon": "^7.3.2",
 		"wct-istanbul": "^0.14.3",
 		"wct-mocha": "^1.0.1",
