+++ test/detach.html
@@ -53,6 +53,7 @@
 
 				const w = imageViewer.detach();
 				if (w == null) {
+					/* eslint-disable-next-line no-console */
 					console.warn('Only gets tested without popup blocker');
 					/* eslint-disable-next-line no-invalid-this */
 					this.skip();
